Swithc tables to uint16

+++ common/enc-internal.h
@@ -4,18 +4,18 @@
 
 namespace Common {
 
-extern const uint32 kWindows1250ConversionTable[128];
-extern const uint32 kWindows1251ConversionTable[128];
-extern const uint32 kWindows1252ConversionTable[128];
-extern const uint32 kWindows1253ConversionTable[128];
-extern const uint32 kWindows1254ConversionTable[128];
-extern const uint32 kWindows1255ConversionTable[128];
-extern const uint32 kWindows1256ConversionTable[128];
-extern const uint32 kWindows1257ConversionTable[128];
-extern const uint32 kMacCentralEuropeConversionTable[128];
-extern const uint32 kLatin1ConversionTable[128];
-extern const uint32 kLatin2ConversionTable[128];
-extern const uint32 kDos850ConversionTable[128];
+extern const uint16 kWindows1250ConversionTable[128];
+extern const uint16 kWindows1251ConversionTable[128];
+extern const uint16 kWindows1252ConversionTable[128];
+extern const uint16 kWindows1253ConversionTable[128];
+extern const uint16 kWindows1254ConversionTable[128];
+extern const uint16 kWindows1255ConversionTable[128];
+extern const uint16 kWindows1256ConversionTable[128];
+extern const uint16 kWindows1257ConversionTable[128];
+extern const uint16 kMacCentralEuropeConversionTable[128];
+extern const uint16 kLatin1ConversionTable[128];
+extern const uint16 kLatin2ConversionTable[128];
+extern const uint16 kDos850ConversionTable[128];
 
 }
